Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,9 +3,9 @@
 
 int main() {
     int a,b,c;
-    float av;
+    double av;
     scanf("%d %d %d",&a,&b,&c);
     av=(a+b+c)/3;
-    printf("Average: %.2f",av);
+    printf("Average: %.2lf",av);
     return 0;
 }
